+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 - Methods ``streamer.Stream.{async_parmap, parmap_async}`` are dropped and merged into ``parmap``.
+- Remove dependency on ``uviloop``.
+- Function ``http.run_app``.
 
 ### Changed